Fixed a discovered issue with L4 positions and rotations. Cleaned up the code a bit
